+++ packages/plugin-node-modules/sources/index.ts
@@ -24,11 +24,12 @@ const plugin: Plugin = {
       default: NodeModulesHoistingLimits.NONE,
     },
     nmMode: {
-      description: `If set to "hardlinks" Yarn will utilize hardlinks to reduce disk space consumption inside node_modules directories`,
+      description: `If set to "hardlinks" Yarn will utilize hardlinks to reduce disk space consumption inside node_modules directories. With "cas" Yarn will use global content addressable storage to reduce node_modules size across projects that use "cas" option.`,
       type: SettingsType.STRING,
       values: [
         NodeModulesMode.CLASSIC,
         NodeModulesMode.HARDLINKS,
+        NodeModulesMode.CAS,
       ],
       default: NodeModulesMode.CLASSIC,
     },