+++ CHANGELOG.md
@@ -56,7 +56,10 @@ yarn set version 3.0.0-rc.1
   - Projects that make use of such dependencies will have to be run with the `--preserve-symlinks` Node option if they wish to access their dependencies.
   - Because Yarn installs will never modify files outside of the project due to security reasons, sub-dependencies of packages with `portal:` must be hoisted outside of the portal. Failing that (for example if the portal package depends on something incompatible with the version hoisted via another package), the linker will produce an error and abandon the install.
 
-- The node-modules linker can now utilize hardlinks for duplicate packages inside `node_modules` to reduce disk space consumption, when `nmMode: hardlinks` setting is used.
+- The node-modules linker can now utilize hardlinks. The new setting `nmMode: classic | hardlinks | cas` specifies which `node_modules` strategy should be used:
+ - `classic` - no hardlinks will be used
+ - `hardlinks` - standard `node_modules` layout will be used with hardlinks inside the project only
+ - `cas` - global content addressable storage will be used across projects having this option
 
 ### Bugfixes